+++ tests/test_ingest.py
@@ -123,8 +123,8 @@ class TestIngestLinks(BaseTestCase):
         data = ingest_links({"STAR/en-US": [tile]}, self.channels[0].id)
         directory_id = data["STAR/en-US"][0]["directoryId"]
 
-        # the biggest ID is 99 - next one should be 100
-        assert_equal(100, directory_id)
+        # the biggest ID is 30 - next one should be 100
+        assert_equal(31, directory_id)
 
     def test_id_not_duplicated(self):
         """
@@ -162,7 +162,7 @@ class TestIngestLinks(BaseTestCase):
         }, self.channels[0].id)
         directory_id_star = data["STAR/en-US"][0]["directoryId"]
         directory_id_ca = data["CA/en-US"][0]["directoryId"]
-        assert_equal(100, directory_id_star)
+        assert_equal(31, directory_id_star)
         assert_not_equal(data["STAR/en-US"][1]["directoryId"], directory_id_star)
         assert_equal(directory_id_ca, directory_id_star)
 
@@ -182,11 +182,11 @@ class TestIngestLinks(BaseTestCase):
 
         data = ingest_links({"STAR/en-US": tiles_star}, self.channels[0].id)
         directory_id = data["STAR/en-US"][0]["directoryId"]
-        assert_equal(100, directory_id)
+        assert_equal(31, directory_id)
 
         data = ingest_links({"STAR/en-US": tiles_star}, self.channels[0].id)
         directory_id = data["STAR/en-US"][0]["directoryId"]
-        assert_equal(100, directory_id)
+        assert_equal(31, directory_id)
 
     def test_error_mid_ingestion(self):
         """